Converted over to using cached objects a little more with ModelStore
Why? Because why not! What started as a simple solution to the problem of circular model dependencies has now spiraled out of control as I decide to solve a non-existent resources problem. But you know what? It feels good to be pure. Also if something goes horribly wrong, this commit is a likely culprit.